[visual only] The One Way Round: show an indicator for going the wrong way.
This is technically a glitch in the map design that seems hard to fix - but also
is not worth fixing. People who go against the arrow AND do not keep their
platform will get softlocked (and have to press escape) though.

Maybe will fix it in a better way later, but for now this will do.