Only skip loading tiles from same neighbor if warpzone states haven't changed.
Otherwise, this is incorrect and can be shown to be so in The Tower Of
Impossible, leading to a glitched entrance when one causes the losing
condition of the room while still viewing the entrance.

After this, changing on-screen warpzone states should be safe.

Note: this change is slightly dangerous and need a full integration test.
Should also consider if this optimization is needed at all or if it can go.