On GitHub Actions, enable software rendering on macOS.
This way, we are verifying that the game works on macOS too, now. Partially,
that is; in a later step, we should try finding a way to also very the app
bundle.

This uses a weird library injection hack, as GLFW explicitly forbids software
rendering, and additionally sets a few flags that Apple's SW renderer does not
support. This game doesn't need any of the extra stuff, however, other games
might, so this fix is not universal.