Apply a better hack for https://github.com/golang/go/issues/57288
Instead of using an artificial format string, I am now accessing the Get
method on the Po object via the Translator interface. Accessing via an
interface makes it impossible for go vet to detect if it's a printf-like
function, thereby evading the false positives due to Get not acting like
printf if it only gets one arg.