Minor "confusion fix".
One could previously, by a very intricate maneuver, bring two platforms through
this section by making a perfectly timed jump that blocks the door with one
platform and grabs the other from the switch.

This maneuver is very tricky, probably TAS-ish, and was not intended - plus,
it gains nothing, but may confuse a player into thinking they have to do this.

Thus, fixed it by some playerclip that makes it impossible to ever get the
platform back shortly after it triggered the switch.

Thanks, polypoyo, for finding it!